Add ResourceViewer views
There are two distinct models of resources graphs being introduced, a
control plane and a data plane. The control plane is only visible for a
Knative Service at the moment and is limited to Service, Configuration,
Revision and Route resources. The data plane is only visible on Service
and Route resources showing the Configurations and Revisions actively
wired receive traffic for the route.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>